moved admin config components to own files
